Add useful git config setup with commit template
Added a git folder that contains intended useful configurations or set
up for git that can be used by and are recommended for Jornaya
developers.  This includes a commit message template and a git config
file to use the commit message template along with setting up some
colorization for git text (branches for git branch or files for git
status) and some recommended Aliases.